refactor: TypeMap object로 typeMap 정의
