+++ feature/calculator/src/main/java/com/hongstudio/feature/calculator/navigation/CalculatorNavigation.kt
@@ -7,14 +7,12 @@ import androidx.navigation.compose.composable
 import androidx.navigation.toRoute
 import com.hongstudio.core.model.CalculatorData
 import com.hongstudio.core.model.CalculatorSelected
-import com.hongstudio.core.navigation.CalculatorSelectedNavType
 import com.hongstudio.core.navigation.MainTabRoute
 import com.hongstudio.core.navigation.Route
-import com.hongstudio.core.navigation.calculatorDataTypeMap
+import com.hongstudio.core.navigation.TypeMap
 import com.hongstudio.feature.calculator.CalculatorInstructionRoute
 import com.hongstudio.feature.calculator.CalculatorResultRoute
 import com.hongstudio.feature.calculator.CalculatorRoute
-import kotlin.reflect.typeOf
 
 fun NavController.navigateCalculator(calculatorSelected: CalculatorSelected) {
     navigate(Route.Calculator(calculatorSelected = calculatorSelected))
@@ -37,7 +35,7 @@ fun NavGraphBuilder.calculatorNavGraph(
     }
 
     composable<Route.Calculator>(
-        typeMap = mapOf(typeOf<CalculatorSelected>() to CalculatorSelectedNavType)
+        typeMap = TypeMap.calculatorSelectedTypeMap
     ) { navBackStackEntry ->
         val calculatorSelected = navBackStackEntry.toRoute<Route.Calculator>().calculatorSelected
         CalculatorRoute(
@@ -48,7 +46,7 @@ fun NavGraphBuilder.calculatorNavGraph(
     }
 
     composable<Route.CalculatorResult>(
-        typeMap = calculatorDataTypeMap
+        typeMap = TypeMap.calculatorDataTypeMap
     ) {
         CalculatorResultRoute(padding = padding)
     }
